Data source updated to SqlServer

+++ Controllers/CustomerController.cs
@@ -62,7 +62,7 @@ namespace mr_mat_api.Controllers
 
         // GET: api/Customer/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Customer>> GetCustomer(long id)
+        public async Task<ActionResult<Customer>> GetCustomer(int id)
         {
             var customer = await _context.Customers.FindAsync(id);
 
@@ -78,7 +78,7 @@ namespace mr_mat_api.Controllers
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see https://aka.ms/RazorPagesCRUD.
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutCustomer(long id, Customer customer)
+        public async Task<IActionResult> PutCustomer(int id, Customer customer)
         {
             if (id != customer.Id)
             {
@@ -120,7 +120,7 @@ namespace mr_mat_api.Controllers
 
         // DELETE: api/Customer/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Customer>> DeleteCustomer(long id)
+        public async Task<ActionResult<Customer>> DeleteCustomer(int id)
         {
             var customer = await _context.Customers.FindAsync(id);
             if (customer == null)
@@ -134,7 +134,7 @@ namespace mr_mat_api.Controllers
             return customer;
         }
 
-        private bool CustomerExists(long id)
+        private bool CustomerExists(int id)
         {
             return _context.Customers.Any(e => e.Id == id);
         }
